+++ src/repositories/configRepository.js
@@ -34,7 +34,8 @@ export default {
         startCalendarYesterday: false,
         notificationIndicator: true,
         autoReorderTasks: false,
-        moveCompletedTaskToBottom: true
+        moveCompletedTaskToBottom: true,
+        fullscreenToDoModal: false
       };
       storageRepository.set("config", default_config);
       return default_config;