+++ changelog.md
@@ -10,6 +10,7 @@ Jun 1, 2023
 - Improving splash screen
 - Adding sent subtask to bottom on clomplete setting
 - Todo Description markdown link now redirect to external link in web browser
+- Adding week start on monday setting
 
 
 ## v2.0.0